Update germany-local-functionality.md
EU Sales List is the same functionality as the VAT Vies. That's why I suggest to move this up to the "VAT section"
And actually there is not only the option to report through VAT Vies (ELMA5). 
The second option is the localized REP 88 - VAT- VIES Declaration Disk: With this REP 88 users can create a csv file which can then be uploaded to the Elster web page. ELMA5 is for those customers who have to report 1000 or more records per period whereas the csv file is the recommended option for smaller customers / less records to be reported in VAT Vies