Update to Search docs (#1264)
* Update ui-enter-criteria-filters.md

We have feedback that having filter criteria in the Search box is confusing. It is after all an advanced feature. We show some examples of using filter criteria in Search, but we don't hint that actually all filter criteria work here. I have added some text to indicate that.

What i could not fix is how prominent the section "Fine-tuning..." is at the top of the document and in the Search section.
I am open to any ways of de-emphasizing this. For example, it could be moved as a subsection at the end of Filtering. Or maybe, reducing the text in this section eg. keep the table of examples but remove the bullet points. Maybe the bullet points and table are better suited in the Filtering section where we talk about filter criteria.

* Update ui-search-filter-faq.md

Added explanation about how Search works under the hood based on ICMs.

* Update ui-search-filter-faq.md

* Update ui-enter-criteria-filters.md

* Update ui-enter-criteria-filters.md

* Update ui-search-filter-faq.md

Co-authored-by: jswymer <jswymer@microsoft.com>