Restore function is related to sales, not purchase
Restore function is only related to sales side. It should be clarified here:
For archived SALES documents where the original still exists and is not posted, you can use the **Restore** function to overwrite the original with the archived version of the document. This is practical if you need to restore the contents of a document to an earlier state.