Federal TAX in Mexico is Value Added Tax, not Sales Tax
It is missleading to name it Sales Tax as this category only applies to USA and CANADA States' and it is different from Sales Tax. A notification should be noted in order for the users to differentiate the two kind of taxes. And remarking that Sales Tax is not a tax requirement in Mexico.