Power BI is not supported on Windows Client
emphasize in the On-Premise deployment that Power BI is not supported on Windows Client. 
The reason behind is that the old component used to display Power BI reports is based on Internet Explorer and it will not be updated accordingly to client runtime team statement and priorities. Partners, customers and users should move to web client where this component is updated and work flawlessly.